Nurodamus-model to install bare sim models
In this refactoring we take into consideration that it might tricky to
maintain 2* models (neurodamus/bare), and that having a single neurodamus that
dlopens the model might result in a fragile setup.

In this approach, neurodamus-model extends sim-model and takes the
required steps to install a bare sim model (libnrnmech / libcorenrnmech)
but proceeds to install the combined neurodamus+model with the advantage
of doing an incremental compilation (all mod2c output and .o files are
reused) both for neuron and coreneuron builds.
The resulting libs (libnrnmech_damus / libcorenrnmech_*_damus) are deployed
besides the original ones, while the new special/special-core replace
the original "bare" specials

Witht this reorganization, the logic was largely broken down and split
between mod compilation (in sim-model) and combining the model with neurodamus
sources and dependencies.